Crate swc_allocator

Allocator for swc.

Features

scoped: Enable scoped mode.

Modes

Default mode

In default mode, [crate::boxed::Box] and [crate::vec::Vec] are identical to the original types in std.

Scoped mode

You need to enable scoped feature to use this mode.

In scoped mode you can use [FastAlloc] to make [crate::boxed::Box] and [crate::vec::Vec] very fast.

In this mode, you need to be careful while using [crate::boxed::Box] and [crate::vec::Vec]. You should ensure that [Allocator] outlives all [crate::boxed::Box] and [crate::vec::Vec] created in the scope.

Recommened way to use this mode is to wrap the whole operations in a call to [Allocator::scope].

Modules

allocators

Various flavors of allocators

api

Various flavors of allocator.

Macros

nightly_only

This expands to the given tokens if the nightly feature is enabled.